:recycle: [Refactor] ui/index.tsx 분리 #main
